go/cmd/roothash: remove unneeded roothash export
